// Server/old-src/multiplayer/types.ts
export type ObjectType = Record<string, unknown> & {
  _id: string;
  _parent?: string;
  _children?: readonly ObjectType[];
};

// /**
//  * @description
//  *  c = create
//  *  u = update
//  *  d = delete
//  */
// export type ChangeObjectType = {
//   c?: ObjectType;
//   u?: ObjectType;
//   d?: string;
// };